% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWS_data.R
\docType{data}
\name{BWS_data}
\alias{BWS_data}
\title{Patients potentially affected by the Beckwith-Wiedemann syndrome
(BWS)}
\format{A data frame consisting of nine independent assays in two
CpG islands located on SDHC gene promoter (SDHC_CpG:17 and SDHC_CpG:27)
and seven imprinted Differentially Methylated Regions (PLAGL1, GRB10, 
MEST, NRAP, H19/IGF2, KCNQ1OT1 and MEG3) with 115 CpG positions and 41 
variables:
\describe{
  \item{Target}{character: DMR/CpG island name}
  \item{Chrom}{integer: Chromosome}
  \item{CpG_pos}{integer: CpG position}
  \item{0.0\%, 25.0\%, 50.0\%, 75.0\%, 100.0\%}{numeric: Actual Methylation Percentages (AMPs)}
  \item{NC15, NC13, NC06, NC05, NC03, NC01, HC37, HC34, HC33, HC32, HC31, HC29, HC25, HC23, HC22}{numeric: Healthy controls}
  \item{B5A42, B5B37, B5C09, B5C06, B5C05, B5C41, B5C04, B5C31, B5C03, B5C28, B5C02, B5C19, B5C17, B5C14, B5C13, B5C12, B5C01, B5B38}{numeric: Patients potentially affected by the Beckwith-Wiedemann
        Syndrome (BWS)}
}}
\usage{
data(BWS_data)
}
\description{
Dataset containing human genomic control DNA (Promega), with five 
distinct actual methylation percentages (0\%, 25\%, 50\%, 75\% and 
100\%) obtained by mixing different ratios of unmethylated and fully
methylated human control DNA, 15 healthy controls DNA and 18 patients
DNA.
}
\examples{
data(BWS_data)
head(BWS_data)
}
\references{
Ochoa E, Zuber V, Fernandez-Jimenez N, Bilbao JR, Clark 
GR, Maher ER, Bottolo L. MethylCal: Bayesian calibration of methylation
levels. Submitted. 2019.
}
\keyword{Beckwith-Wiedemann}
\keyword{dataset}
\keyword{syndrome}
