% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploratoryPlot.R
\name{ExploratoryPlot}
\alias{ExploratoryPlot}
\title{Preliminary data exploration}
\usage{
ExploratoryPlot(data, Target = NULL, opt_BoxPlot = 0, dir = NULL,
  cex_par = 1.25)
}
\arguments{
\item{data}{Formatted input data frame obtained from the function
\code{\link{Formatting}}.}

\item{Target}{Name of the target DMR/CpG island/gene to be visualised.}

\item{opt_BoxPlot}{Boxplot option: If \code{opt_BoxPlot = 0} (default),
BoxPlot is centered around the median, whereas if \code{opt_BoxPlot = 1}
it is centered around the mean.}

\item{dir}{In Unix-specific OS, user-specified directory where the
plots in \code{\link[grDevices]{pdf}} format are saved. If the directory
is not specified, figures are saved in the current working directory.}

\item{cex_par}{Number indicating the amount by which plotting text
and symbols should be scaled relative to the default (\code{cex_par = 1}).}
}
\value{
This function returns five plots. The first scatterplot depicts
the values of the recorded apparent methylation levels at each Actual
Methylation Percentage (AMP). The second one summirises the recorded
apparent methylation levels at each AMP by a BoxPlot, fitting a cubic
polynomial regression (Moskalev et al., 2011) on the median (mean)
of the data. The third and the four plots show the apparent methylation
percentage after PCR at consecutive CpGs and at each CpG physical
position on the chromosome. Finally, the fifth plot presents the 
apparent methylation levels at consecutive CpGs stratified by AMPs.

In Unix-specific OS, figures are saved in the current directory,
unless otherwise specified by the user, in \code{\link[grDevices]{pdf}}
format. In Windows OS, figures are printed on the screen.
}
\description{
Data visualisation of the standard control experiment.
}
\examples{
data(BWS_data)
AMP = c(0, 25, 50, 75, 100)
data = Formatting(BWS_data, AMP = AMP)
ExploratoryPlot(data, Target = "KCNQ1OT1")

data(Celiac_data)
AMP = c(0, 12.5, 25, 37.5, 50, 62.5, 87.5, 100)
data = Formatting(Celiac_data, AMP = AMP)
ExploratoryPlot(data, Target = "NFKBIA", opt_BoxPlot = 1)
}
\references{
Ochoa E, Zuber V, Fernandez-Jimenez N, Bilbao JR, Clark
GR, Maher ER and Bottolo L. MethylCal: Bayesian calibration of methylation
levels. Submitted. 2019.

Moskalev EA, Zavgorodnij MG, Majorova SP, Vorobjev IA,
Jandaghi P, Bure IV, Hoheisel JD. Correction of PCR-bias in quantitative
DNA methylation studies by means of cubic polynomial regression.
Nucleic Acids Res. 2011; 39(11):e77. (\href{https://www.ncbi.nlm.nih.gov/pubmed/21486748}{PubMed})
}
\keyword{Data}
\keyword{control}
\keyword{data}
\keyword{experiment}
\keyword{exploration,}
\keyword{standard}
\keyword{visualisation,}
