% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatting.R
\name{Formatting}
\alias{Formatting}
\title{Reformatting utility function}
\usage{
Formatting(data, AMP = NULL, n_Control = 0, n_Case = 0)
}
\arguments{
\item{data}{Input data frame formatted. It should contain the
following columns: "Target" (name of the target DMR/CpG island/gene 
name), "Chrom" (Chromosome), "CpG pos" (CpG position) along with
the columns for the measurements of the apparent level of 
methylation for each Actual Methylation Percentage (AMP) and for 
each CpG. Missing values are allowed and should be marked as NA.
This utility function can process several Targets if included in
the input data frame.}

\item{AMP}{Numeric vector of Actual Methylation Percentages (AMPs).}

\item{n_Control}{Number of controls samples.}

\item{n_Case}{Number of case samples.}
}
\value{
Long formatted output data frame. It includes the name of 
the DRM/CpG island/gene ("Target"), chromosome ("Chrom"), CpG position
("CpG_group"), at which CpG the measurement has been taken ("CpG_group")
within each DRM/CpG island/gene, a label attached to each CpG 
("CpG_group_label"), at which AMP the measurement has been taken 
("AMP_group"), a label attached to each CpG ("CpG_group_label")
as well as the apparent ("y") and actual ("x") methylation levels.

If control/cases samples are included, methylation levels are stored
in the long formatted data frame at each row corresponding to 
CpG_group = 1 and AMP_group = 1.
}
\description{
Reformatting data frame for MethylCal analysis.
}
\examples{
data(BWS_data)
AMP = c(0, 25, 50, 75, 100)
data = Formatting(BWS_data, AMP = AMP)
data = Formatting(BWS_data, AMP = AMP, n_Control = 15)
data = Formatting(BWS_data, AMP = AMP, n_Control = 15, n_Case = 18)

data(Celiac_data)
AMP = c(0, 12.5, 25, 37.5, 50, 62.5, 87.5, 100)
data = Formatting(Celiac_data, AMP = AMP)
data = Formatting(Celiac_data, AMP = AMP, n_Control = 13)
data = Formatting(Celiac_data, AMP = AMP, n_Control = 13, n_Case = (2 * 17))
}
\references{
Ochoa E, Zuber V, Fernandez-Jimenez N, Bilbao JR, Clark 
GR, Maher ER, Bottolo L. MethylCal: Bayesian calibration of methylation
levels. Submitted. 2019.
}
\keyword{Reformatting,}
\keyword{data}
\keyword{frame}
\keyword{input}
