% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylCalCalibrationPlot.R
\name{MethylCalCalibrationPlot}
\alias{MethylCalCalibrationPlot}
\title{MethylCal calibration plots}
\usage{
MethylCalCalibrationPlot(data, Target = NULL, prior = "LG",
  level = 0.95, dir = NULL, printing = TRUE, cex_par = 1.25)
}
\arguments{
\item{data}{Formatted input data frame obtained from the function
\code{\link{Formatting}}.}

\item{Target}{Name of the target DMR/CpG island/gene to be visualised.}

\item{prior}{Prior distribution set-up for the random effects and
the Latent Gaussian Field (Rue et al., 2009). Three different priors
are implemented:
\itemize{
  \item \code{LG}: log-Gamma prior is the default prior with 
  \code{a = 1} and \code{b = 0.1} parametrization;
  \item \code{PC}: Penalized Complexity prior (Simpson et al., 
  2017);
  \item \code{HC}: Half-Cauchy prior (Wang X et al., 2018).
}}

\item{level}{Level of the posterior predictive region.}

\item{dir}{In Unix-specific OS, user-specified directory where the
plots in \code{\link[grDevices]{pdf}} format are saved. If the directory
is not specified, figures are saved in the current working directory.}

\item{printing}{If \code{printing = TRUE} (default), messages are 
printed on the screen regarding the estimated models (mlik = marginal
likelihood, DIC = Deviance Information Criteria, RSS = Residual
Sum of Squares) and the correction of the apparent methylation levels
(Ochoa et al., 2019).}

\item{cex_par}{Number indicating the amount by which plotting text
and symbols should be scaled relative to the default (\code{cex_par = 1}).}
}
\value{
This function returns three plots. The first scatterplot 
depicts the values of the recorded apparent methylation levels at
each Actual Methylation Percentage (AMP) with superimposed MethylCal's
calibration curve (Ochoa et al., 2019) for each CpG (red dashed
line). The second plot presents the apparent methylation levels at 
consecutive CpGs stratified by AMPs with superimposed the predicted
values (red dashed line) as well as the (1-level)\% posterior predictive
region (dashed-dotted red lines). Finally, the third plot is the 
scatterplot of the corrected actual methylation percentage at each
AMP for all CpGs within a DRM/CpG island/gene.

In Unix-specific OS, figures are saved in the current directory,
unless otherwise specified by the user, in \code{\link{pdf}} format.
In Windows OS, figures are printed on the screen.
}
\description{
Visualisation of MethylCal calibration of the standard control experiment.
}
\examples{
data(BWS_data)
AMP = c(0, 25, 50, 75, 100)
data = Formatting(BWS_data, AMP = AMP)
MethylCalCalibrationPlot(data, Target = "KCNQ1OT1", prior = "HC")

data(Celiac_data)
AMP = c(0, 12.5, 25, 37.5, 50, 62.5, 87.5, 100)
data = Formatting(Celiac_data, AMP = AMP)
MethylCalCalibrationPlot(data, Target = "NFKBIA", level = 0.99, printing = FALSE)
}
\references{
Ochoa E, Zuber V, Fernandez-Jimenez N, Bilbao JR, Clark 
GR, Maher ER and Bottolo L. MethylCal: Bayesian calibration of methylation
levels. Submitted. 2019.

Wang X, Ryan YY, Faraway JJ. Bayesian Regression Modeling
with INLA. 2018, 1st edition. Chapman and Hall/CRC.

Simpson S, Rue H, Riebler A, Martins TG, Sorbye SH. Penalising
model component complexity: A principled, practical approach to constructing
priors. Statist Sci. 2017; 1:1-28.  (\href{https://doi.org/10.1214/16-STS576}{doi})

Rue H, Martino S, Chopin N. Approximate Bayesian inference
for latent Gaussian models by using integrated nested Laplace approximations.
J Roy Stat Soc B Met. 2009; 71(2):319-392. (\href{https://doi.org/10.1111/j.1467-9868.2008.00700.x}{doi})
}
\keyword{Data}
\keyword{MethylCal}
\keyword{calibration,}
\keyword{control}
\keyword{experiment}
\keyword{standard}
\keyword{visualisation,}
