% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylCalCorrection.R
\name{MethylCalCorrection}
\alias{MethylCalCorrection}
\title{MethylCal correction of case/controls samples}
\usage{
MethylCalCorrection(data, Target = NULL, prior = "LG", n_Control = 0,
  n_Case = 0, level_Control = 0.9986501, opt_BoxPlot = 0,
  dir = NULL, printing = TRUE, plotting = TRUE, cex_par = 1.25)
}
\arguments{
\item{data}{Formatted input data frame obtained from the function
\code{\link{Formatting}}.}

\item{Target}{Name of the target DMR/CpG island/gene to be visualised.}

\item{prior}{Prior distribution set-up for the random effects and 
the Latent Gaussian Field (Rue et al., 2009). Three different priors
are implemented:
\itemize{
  \item \code{LG}: log-Gamma prior is the default prior with 
  \code{a = 1} and \code{b = 0.1} parametrization;
  \item \code{PC}: Penalized Complexity prior (Simpson et al., 
  2017);
  \item \code{HC}: Half-Cauchy prior (Wang X et al., 2018).
}}

\item{n_Control}{Number of controls samples.}

\item{n_Case}{Number of case samples.}

\item{level_Control}{Level of significance of the differential
methylation test between case and control corrected samples. Default
value is \code{level_Control = 0.9986501} which correspond to z = 3
quantile for normally distributed data.}

\item{opt_BoxPlot}{Boxplot option: If \code{opt_BoxPlot = 0} (default), 
BoxPlot is centered around the median, whereas if \code{opt_BoxPlot = 1} 
it is centered around the mean.}

\item{dir}{In Unix-specific OS, user-specified directory where the
plots in \code{\link[grDevices]{pdf}} format are saved. If the directory
is not specified, figures are saved in the current working directory.}

\item{printing}{If \code{printing = TRUE} (default), the corrected 
methylation levels of the case/control samples using MethylCal calibration
(Ochoa et al., 2019) are printed on the screen.}

\item{plotting}{If \code{plotting = TRUE} (default), the corrected 
methylation levels for the control samples as well as its (1-\code{level_Control)}\%
confidence interval are depicted. BoxPlot of the corrected methylation
levels are also shown for each control sample. If cases are
included, a second figure presents a BoxPlot for each case sample
as well as the controls' (1-\code{level_Control)}\% confidence 
interval and the hyper- and hypo-methylated cases (top red triangles).}

\item{cex_par}{Number indicating the amount by which plotting text
and symbols should be scaled relative to the default (\code{cex_par = 1}).}
}
\value{
This function returns the corrected methylation level for
the control and (if selected) case samples using MethylCal calibration.
Based a parametric t-test at (1-\code{level_Control)}\%, hyper- and
hypo-methylated cases are also flagged.

In Unix-specific OS, figures are saved in the current directory,
unless otherwise specified by the user, in \code{\link{pdf}} format.
In Windows OS, figures are printed on the screen.
}
\description{
Correction of case/controls samples using MethylCal calibration
}
\examples{
data(BWS_data)
AMP = c(0, 25, 50, 75, 100)
data = Formatting(BWS_data, AMP = AMP, n_Control = 15)
corr_data = MethylCalCorrection(data, Target = "KCNQ1OT1", n_Control = 15)

data(Celiac_data)
AMP = c(0, 12.5, 25, 37.5, 50, 62.5, 87.5, 100)
data = Formatting(Celiac_data, AMP = AMP, n_Control = 13, n_Case = (2 * 17))
corr_data = MethylCalCorrection(data, Target = "NFKBIA", n_Control = 13, n_Case = (2 * 17), 
opt_BoxPlot = 1)
}
\references{
Ochoa E, Zuber V, Fernandez-Jimenez N, Bilbao JR, Clark 
GR, Maher ER and Bottolo L. MethylCal: Bayesian calibration of methylation
levels. Submitted. 2019.

Wang X, Ryan YY, Faraway JJ. Bayesian Regression Modeling
with INLA. 2018, 1st edition. Chapman and Hall/CRC.

Simpson S, Rue H, Riebler A, Martins TG, Sorbye SH. Penalising
model component complexity: A principled, practical approach to constructing
priors. Statist Sci. 2017; 1:1-28.  (\href{https://doi.org/10.1214/16-STS576}{doi})

Rue H, Martino S, Chopin N. Approximate Bayesian inference
for latent Gaussian models by using integrated nested Laplace approximations.
J Roy Stat Soc B Met. 2009; 71(2):319-392. (\href{https://doi.org/10.1111/j.1467-9868.2008.00700.x}{doi})
}
\keyword{Case}
\keyword{MethylCal}
\keyword{and}
\keyword{calibration}
\keyword{control}
\keyword{samples,}
