% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoskalevCorrection.R
\name{MoskalevCorrection}
\alias{MoskalevCorrection}
\title{Moskalev correction of case/controls samples}
\usage{
MoskalevCorrection(data, Target = NULL, n_Control = 0, n_Case = 0,
  level_Control = 0.9986501, opt_BoxPlot = 0, dir = NULL,
  printing = TRUE, plotting = TRUE, cex_par = 1.25)
}
\arguments{
\item{data}{Formatted input data frame obtained from the function
\code{\link{Formatting}}.}

\item{Target}{Name of the target DMR/CpG island/gene to be visualised.}

\item{n_Control}{Number of controls samples.}

\item{n_Case}{Number of case samples.}

\item{level_Control}{Level of significance of the differential
methylation test between case and control corrected samples. Default
value is \code{level_Control = 0.9986501} which correspond to z = 3
quantile for normally distributed data.}

\item{opt_BoxPlot}{Boxplot option: If \code{opt_BoxPlot = 0} (default), 
BoxPlot is centered on the median, whereas if \code{opt_BoxPlot = 1} 
it is centered on the mean.}

\item{dir}{In Unix-specific OS, user-specified directory where the
plots in \code{\link[grDevices]{pdf}} format are saved. If the directory
is not specified, figures are saved in the current working directory.}

\item{printing}{If \code{printing = TRUE} (default), the corrected 
methylation levels of the case/control samples using Moskalev's 
cubic polynomial regression (Moskalev et al., 2011) are printed on
the screen.}

\item{plotting}{If \code{plotting = TRUE} (default), the corrected 
methylation levels for the control samples as well as its (1-\code{level_Control)}\%
confidence interval are depicted. BoxPlot of the corrected methylation
levels are also shown for each control sample. If cases are included,
a second figure presents a BoxPlot for each case sample as well as
the controls' (1-\code{level_Control)}\% confidence interval and 
the hyper- and hypo-methylated cases (top red triangles).}

\item{cex_par}{Number indicating the amount by which plotting text
and symbols should be scaled relative to the default (\code{cex_par = 1}).}
}
\value{
This function returns the corrected methylation level for
the control and (if selected) case samples using Moskalev's cubic
polynomial regression. Based a parametric t-test at (1-\code{level_Control)}\%,
hyper- and hypo-methylated cases are also flagged.

In Unix-specific OS, figures are saved in the current directory,
unless otherwise specified by the user, in \code{\link{pdf}} format.
In Windows OS, figures are printed on the screen.
}
\description{
Correction of case/controls samples by using Moskalev's cubic polynomial
regression.
}
\examples{
data(BWS_data)
AMP = c(0, 25, 50, 75, 100)
data = Formatting(BWS_data, AMP = AMP, n_Control = 15)
corr_data = MoskalevCorrection(data, Target = "KCNQ1OT1", n_Control = 15)

data(Celiac_data)
AMP = c(0, 12.5, 25, 37.5, 50, 62.5, 87.5, 100)
data = Formatting(Celiac_data, AMP = AMP, n_Control = 13, n_Case = (2 * 17))
corr_data = MoskalevCorrection(data, Target = "NFKBIA", n_Control = 13, n_Case = (2 * 17), 
opt_BoxPlot = 1)
}
\references{
Ochoa E, Zuber V, Fernandez-Jimenez N, Bilbao JR, Clark 
GR, Maher ER and Bottolo L. MethylCal: Bayesian calibration of methylation
levels. Submitted. 2019.

Moskalev EA, Zavgorodnij MG, Majorova SP, Vorobjev IA, 
Jandaghi P, Bure IV, Hoheisel JD. Correction of PCR-bias in quantitative
DNA methylation studies by means of cubic polynomial regression. 
Nucleic Acids Res. 2011; 39(11):e77. (\href{https://www.ncbi.nlm.nih.gov/pubmed/21486748}{PubMed})
}
\keyword{Case}
\keyword{Moskalev}
\keyword{and}
\keyword{calibration}
\keyword{control}
\keyword{samples,}
