% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WarneckeCalibrationPlot.R
\name{WarneckeCalibrationPlot}
\alias{WarneckeCalibrationPlot}
\title{Warnecke calibration plots}
\usage{
WarneckeCalibrationPlot(data, Target = NULL, level = 0.95,
  dir = NULL, cex_par = 1.25)
}
\arguments{
\item{data}{Formatted input data frame obtained from the function
\code{\link{Formatting}}.}

\item{Target}{Name of the target DMR/CpG island/gene to be visualised.}

\item{level}{Level of significance of prediction interval obtained with
Warnecke's calibration curve.}

\item{dir}{In Unix-specific OS, user-specified directory where the
plots in \code{\link[grDevices]{pdf}} format are saved. If the directory
is not specified, figures are saved in the current working directory.}

\item{cex_par}{Number indicating the amount by which plotting text
and symbols should be scaled relative to the default (\code{cex_par = 1}).}
}
\value{
This function returns four plots. The first scatterplot depicts
the values of the recorded apparent methylation levels at each Actual
Methylation Percentage (AMP) with superimposed Warnecke's calibration
curve (Warnecke et al., 1997) for each CpG (red dashed line). The
second plot presents the apparent methylation levels at consecutive
CpGs stratified by AMPs with superimposed the (1-level)\% prediction
interval (dashed-dotted red lines). The third plot is the scatterplot
of the corrected methylation levels at each AMP for all CpGs within
a DRM/CpG island/gene. Finally, the fourth plot is the hyperbolic
coefficient (Moskalev et al., 2011) of the corrected methylation 
levels regressed on the AMPs for each CpG.

In Unix-specific OS, figures are saved in the current directory,
unless otherwise specified by the user, in \code{\link{pdf}} format.
In Windows OS, figures are printed on the screen.
}
\description{
Visualisation of Warnecke calibration of the standard control experiment.
}
\examples{
data(BWS_data)
AMP = c(0, 25, 50, 75, 100)
data = Formatting(BWS_data, AMP = AMP)
WarneckeCalibrationPlot(data, Target = "KCNQ1OT1")

data(Celiac_data)
AMP = c(0, 12.5, 25, 37.5, 50, 62.5, 87.5, 100)
data = Formatting(Celiac_data, AMP = AMP)
WarneckeCalibrationPlot(data, Target = "NFKBIA", level = 0.99)
}
\references{
Ochoa E, Zuber V, Fernandez-Jimenez N, Bilbao JR, Clark 
GR, Maher ER and Bottolo L. MethylCal: Bayesian calibration of methylation
levels. Submitted. 2019.

Moskalev EA, Zavgorodnij MG, Majorova SP, Vorobjev IA, 
Jandaghi P, Bure IV, Hoheisel JD. Correction of PCR-bias in quantitative
DNA methylation studies by means of cubic polynomial regression. 
Nucleic Acids Res. 2011; 39(11):e77. (\href{https://www.ncbi.nlm.nih.gov/pubmed/21486748}{PubMed})

Warnecke PM, Stirzaker C, Melki JR, Millar DS, Paul CL, 
Clark SJ. Detection and measurement of PCR bias in quantitative 
methylation analysis of bisulphite-treated DNA. Nucleic Acids Res. 
1997; 25(21):4422-6. (\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC147052/}{PMC})
}
\keyword{Data}
\keyword{Warnecke}
\keyword{calibration,}
\keyword{control}
\keyword{experiment}
\keyword{standard}
\keyword{visualisation,}
